+++ opensyde_syde_coder_c/libs/osy_core/project/system/node/C_OscNodeDataPoolFiler.cpp
@@ -650,8 +650,8 @@ int32_t C_OscNodeDataPoolFiler::h_LoadDataPoolLists(std::vector<C_OscNodeDataPoo
       if (u32_ExpectedSize != orc_NodeDataPoolLists.size())
       {
          C_SclString c_Tmp;
-         c_Tmp.PrintFormatted("Unexpected list count, expected: %i, got %i", u32_ExpectedSize,
-                              orc_NodeDataPoolLists.size());
+         c_Tmp.PrintFormatted("Unexpected list count, expected: %u, got %u", u32_ExpectedSize,
+                              static_cast<uint32_t>(orc_NodeDataPoolLists.size()));
          osc_write_log_warning("Load file", c_Tmp.c_str());
       }
    }
@@ -745,8 +745,8 @@ int32_t C_OscNodeDataPoolFiler::h_LoadDataPoolListElements(
       if (u32_ExpectedSize != orc_NodeDataPoolListElements.size())
       {
          C_SclString c_Tmp;
-         c_Tmp.PrintFormatted("Unexpected data element count, expected: %i, got %i", u32_ExpectedSize,
-                              orc_NodeDataPoolListElements.size());
+         c_Tmp.PrintFormatted("Unexpected data element count, expected: %u, got %u", u32_ExpectedSize,
+                              static_cast<uint32_t>(orc_NodeDataPoolListElements.size()));
          osc_write_log_warning("Load file", c_Tmp.c_str());
       }
    }
